Account for latest changes
