+++ api/src/main/java/com/convallyria/forcepack/api/ForcePackAPI.java
@@ -3,9 +3,6 @@ package com.convallyria.forcepack.api;
 import com.convallyria.forcepack.api.resourcepack.ResourcePack;
 import com.convallyria.forcepack.api.schedule.PlatformScheduler;
 
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 
@@ -31,50 +28,4 @@ public interface ForcePackAPI {
      */
     boolean exemptNextResourcePackSend(UUID uuid);
 
-    /**
-     * Gets whether the specified URL is a default hosted one.
-     * @param url the url to check
-     * @return true if site is a default host
-     */
-    default boolean isDefaultHost(String url) {
-        List<String> warnForHost = List.of("convallyria.com");
-        for (String host : warnForHost) {
-            if (url.contains(host)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    /**
-     * Gets, if there is one, the blacklisted site within a URL
-     * @param url the url to check
-     * @return an {@link Optional<String>} possibly containing the URL of the blacklisted site found
-     */
-    default Optional<String> getBlacklistedSite(String url) {
-        List<String> blacklisted = List.of("mediafire.com");
-        for (String blacklistedSite : blacklisted) {
-            if (url.contains(blacklistedSite)) {
-                return Optional.of(url);
-            }
-        }
-        return Optional.empty();
-    }
-
-    /**
-     * Gets whether the specified URL has a valid ending
-     * @param url the url to check
-     * @return true if URL ends with a valid extension
-     */
-    default boolean isValidEnding(String url) {
-        List<String> validUrlEndings = Arrays.asList(".zip", "dl=1");
-        boolean hasEnding = false;
-        for (String validUrlEnding : validUrlEndings) {
-            if (url.endsWith(validUrlEnding)) {
-                hasEnding = true;
-                break;
-            }
-        }
-        return hasEnding;
-    }
 }
