+++ spigot/src/main/java/com/convallyria/forcepack/spigot/ForcePackSpigot.java
@@ -1,8 +1,8 @@
 package com.convallyria.forcepack.spigot;
 
 import com.convallyria.forcepack.api.ForcePackAPI;
+import com.convallyria.forcepack.api.ForcePackPlatform;
 import com.convallyria.forcepack.api.player.ForcePackPlayer;
-import com.convallyria.forcepack.api.resourcepack.PackFormatResolver;
 import com.convallyria.forcepack.api.resourcepack.ResourcePack;
 import com.convallyria.forcepack.api.resourcepack.ResourcePackVersion;
 import com.convallyria.forcepack.api.schedule.PlatformScheduler;
@@ -59,7 +59,7 @@ import java.util.function.Consumer;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
 
-public final class ForcePackSpigot extends JavaPlugin implements ForcePackAPI {
+public final class ForcePackSpigot extends JavaPlugin implements ForcePackPlatform {
 
     private Translator translator;
     private PlatformScheduler<?> scheduler;
@@ -79,41 +79,7 @@ public final class ForcePackSpigot extends JavaPlugin implements ForcePackAPI {
 
     public Set<ResourcePack> getPacksForVersion(Player player) {
         final int protocolVersion = ProtocolUtil.getProtocolVersion(player);
-        final int packFormat = PackFormatResolver.getPackFormat(protocolVersion);
-
-        log("Searching for a resource pack with pack version " + packFormat);
-
-        ResourcePack anyVersionPack = null;
-        Set<ResourcePack> validPacks = new HashSet<>();
-        for (ResourcePack resourcePack : getResourcePacks()) {
-            final Optional<ResourcePackVersion> version = resourcePack.getVersion();
-            log("Trying resource pack " + resourcePack.getURL() + " (" + (version.isEmpty() ? version.toString() : version.get().toString()) + ")");
-
-            if (version.isEmpty()) {
-                if (anyVersionPack == null) anyVersionPack = resourcePack; // Pick first all-version resource pack
-                validPacks.add(resourcePack); // This is still a valid pack that we want to apply.
-                continue;
-            }
-
-            if (version.get().inVersion(packFormat)) {
-                validPacks.add(resourcePack);
-                log("Added resource pack " + resourcePack.getURL());
-                if (protocolVersion < 765) { // If < 1.20.3, only one pack can be applied.
-                    break;
-                }
-            }
-        }
-
-        if (!validPacks.isEmpty()) {
-            log("Found multiple valid resource packs (" + validPacks.size() + ")");
-            for (ResourcePack validPack : validPacks) {
-                log("Chosen resource pack " + validPack.getURL());
-            }
-            return validPacks;
-        }
-
-        log("Chosen resource pack is " + (anyVersionPack == null ? "null" : anyVersionPack.getURL()));
-        return anyVersionPack == null ? Set.of() : Set.of(anyVersionPack);
+        return getPacksForVersion(protocolVersion);
     }
 
     @Override
@@ -301,28 +267,6 @@ public final class ForcePackSpigot extends JavaPlugin implements ForcePackAPI {
         if (!success) {
             getLogger().severe("Unable to load all resource packs correctly.");
         }
-    }
-
-    private ResourcePackVersion getVersionFromId(String versionId) {
-        if (versionId.equals("all")) {
-            return null;
-        }
-
-        try {
-            // One version?
-            final int fixedVersion = Integer.parseInt(versionId);
-            return ResourcePackVersion.of(fixedVersion, fixedVersion);
-        } catch (NumberFormatException ignored) {
-            try {
-                // Version range?
-                final String[] ranged = versionId.split("-");
-                final int min = Integer.parseInt(ranged[0]);
-                final int max = Integer.parseInt(ranged[1]);
-                return ResourcePackVersion.of(min, max);
-            } catch (NumberFormatException | IndexOutOfBoundsException ignored2) {}
-        }
-
-        throw new IllegalArgumentException("Invalid version id: " + versionId);
     }
 
     private boolean checkPack(@Nullable ResourcePackVersion version, String url, boolean generateHash, @Nullable String hash) {
@@ -514,8 +458,9 @@ public final class ForcePackSpigot extends JavaPlugin implements ForcePackAPI {
         return getConfig().getBoolean("Server.debug");
     }
 
-    public void log(String info) {
-        if (debug()) getLogger().info(info);
+    @Override
+    public void log(String info, Object... format) {
+        if (debug()) getLogger().info(String.format(info, format));
     }
 
     public static ForcePackAPI getAPI() {
